Fixes unit tests for `UISearchBar`.
